refactor(core/forms): Remove some cached state from MapEditor Move MapPair component to its own file
